Updated CSS Media Features to phantom types.
The entire Css.Media module has been updated to work with Phantom Types. This requires testing in the future, but so far it compiles fine. It also requires more documentation refinements in a future commit.

----------

Changed certain input conventions
Where a CSS Media Feature only accepts an integer value that doesn't have a prefixed
character indicating units and those units aren't used elsewhere, I've decided to remove the unit functions that were there and
make the media feature simply accept an Int instead.

This is to make the code smoother and less redundant.

- Removed bits value from color depth media features.
- Removed ratio value from ratio media features.

----------

A bit of renaming
- Renamed 'canHover' to 'hover_' to match naming conventions in Css.
- Renamed 'PointerDevice' to 'PointingAccuracy'.

----------

Documentation improvements
This is incomplete. It passes the bare minimum standards for Elm packages, but not the standards for the phantom-types fork.

I've got to do more in a future commit.

- Where possible from a quick scan of all the functions, documentation has been moved to MDN where possible.
- Tried to write the language of the documentation to be easier to read.
- All exposed types and functions have some sort of documentation.
- Made code examples more accurate (almost all used 'media' instead of 'withMedia').
- Removed lines that mention what keywords a media feature accepts because with phantom types, this is now relatively easy to check.

----------

Experimental media features removed
Mirroring my approach with experimental CSS properties.
- Removed the 'update' CSS Media Feature.

-----------

Other
- Moved phantom type-based Resolution values from Css to Css.Media.