-Many property function fallbacks
The -Many functions now (mostly) fall back to unset when the lists are empty. This is to ensure it's impossible to create CSS properties with empty values and also enable the smoothest possible CSS writing experience.
The one exception is boxShadowMany, as it already falls back to none and I don't know enough about it to mess with it atm.

I also fixed some incorrect documentation relating to -Many functions in this module.