Testing: Revamping existing property tests
I didn't really notice there was a much more comprehensive testing system inside the phantom-types fork, so I have gotten that out, documented it, and split it up between modules for better readability.
Most of it is done, but not all. Grid needs to be finished, as do remaining testing in Phantom.

- Removed value testing as this better testing system has made it obsolete.